pyproject.toml
[project]
name = "ecm-webdev-backend"
version = "0.1.0"
description = "Web development backend course at ECM"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.118.0",
    "flexmock>=0.12.2",
    "httpx>=0.28.1",
    "nanoid>=2.0.0",
    "pyhumps>=3.8.0",
    "pyjwt>=2.10.1",
    "pytest>=8.4.2",
    "python-dotenv>=1.1.1",
    "sqlmodel>=0.0.25",
    "time-machine>=2.19.0",
    "uvicorn>=0.37.0",
    "websocket>=0.2.1",
    "websockets>=15.0.1",
]

[dependency-groups]
dev = [
    "coverage>=7.10.7",
    "flexmock>=0.12.2",
    "mypy>=1.18.2",
    "pytest-asyncio>=1.2.0",
    "syrupy>=5.0.0",
    "time-machine>=2.19.0",
    "types-nanoid>=2.0.0.20240601",
]
